event_summary blueprint: adjust the device name sanitization
The device name sanitize code was changed from "remove all non-conforming characters and then convert some special ones to '_'" to "convert some special characters to '_' and then remove all other non-conforming characters.

In the new version, both of these operations were changed to use regex replace only.  Note that in the "special characters" expression `[' -]` the '-' does not need to be escaped because it is the last item in the regex character class.  The "non-conforming characters" regex was changed to `[^a-z0-9_]`, which reflects that the special characters no longer need to be included.

An important addition to the "special characters" is the single-quote, which was previously omitted.