blueprint: re-implement cooldown with singlemode execution and a delay
Force the automation to use single mode with suppressed warnings (see
https://www.home-assistant.io/docs/automation/modes/).  To implement
the cooldown, add a delay as the last action in the actions.  Also
remove the cooldown from the automation conditions.

The main difference here is that no new triggers will arrive while the
automation is running.  The condition is smaller now, and is limited
to the content of the automation, and not managing how the automation
is run.  With this narrowing, it opens up the possibility of adding
other conditions to the blueprint config in the future.