blueprint: add a filter for frigate object detection
The frigate payload includes a label which indicates what the detected
object is.  This can be useful to separate llm-vision automations to
react differently for different objects.

For example, if a bird is detected, the ai instructions could include
something like, "if you see a bird, try to identify what species it is".
This discription doesn't make much sense if the detected object is a
person, and including separate instructions to the ai makes it more
confusing on both the human and ai side of things.  Allowing for
separate instructions by object will make for easier to understand
automations.