blueprint: add more notification control
When the automation is triggered, a notification is sent immediately to
the user, and then the data is forwarded on to the llm engine.  Once the
llm responds, the previous notification is updated with the description
from the llm.

However, each of these will trigger a push notification on the user's
device, which may ring the notificaiton tone or vibrate.  This may be
undesirable for all users, so a new control is added to the automation
blueprint to suppress the immediate notification.

Additionally, some responses from the llm may not be interesting to the
user, and so a new 'filter' option is added to the automation blueprint
which will suppress the second notification.